refactor(x/tally): refactor error handling in filter and tally
